% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentationChunks}
\alias{documentationChunks}
\title{Read documentation chunks}
\usage{
documentationChunks()
}
\value{
List of chunks of the OMOP documentation.
}
\description{
Read documentation chunks
}
\examples{
\donttest{
documentationChunks()
}

}
