% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationLinks.R
\name{documentationLinks}
\alias{documentationLinks}
\title{List the documentation links to train the OmopAssistant model}
\usage{
documentationLinks()
}
\value{
Named character vector with documentation links.
}
\description{
List the documentation links to train the OmopAssistant model
}
\examples{
documentationLinks()

}
