% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{downloadedModels}
\alias{downloadedModels}
\title{Title}
\usage{
downloadedModels()
}
\value{
Models that have been downloaded.
}
\description{
Title
}
\examples{
\donttest{
library(OmopAssistant)

path <- file.path(tempdir(), "OMOP_DATA_FOLDER")
dir.create(path = path)
Sys.setenv("OMOP_DATA_FOLDER" = path)

downloadedModels()
downloadModel(name = "omop_assistant")
downloadedModels()
}

}
