% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{downloadModel}
\alias{downloadModel}
\title{Download a preexisting model}
\usage{
downloadModel(name = "omop_assistant", overwrite = FALSE)
}
\arguments{
\item{name}{Name of the model to download. See options using \code{avialableModels()}.}

\item{overwrite}{Whether to overwrite a preexisting model.}
}
\value{
Path to the downloaded model.
}
\description{
Download a preexisting model
}
\examples{
\donttest{
library(OmopAssistant)

Sys.setenv("OMOP_DATA_FOLDER" = tempdir())

downloadedModels()
downloadModel(name = "omop_assistant")
downloadedModels()
}

}
