% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pathModel}
\alias{pathModel}
\title{Path to a store}
\usage{
pathModel(name = NULL)
}
\arguments{
\item{name}{Store name.}
}
\value{
Path to the model.
}
\description{
Path to a store
}
\examples{
library(OmopAssistant)

Sys.setenv("OMOP_DATA_FOLDER" = tempdir())
pathModel(name = NULL)

}
