% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{storeCreate}
\alias{storeCreate}
\title{Create a store}
\usage{
storeCreate(embed, name = "omop_assistant", overwrite = FALSE)
}
\arguments{
\item{embed}{A function that is called with a character vector and returns a
matrix of embeddings. Note this function will be serialised and then
deserialised in new R sessions, so it cannot reference to any objects in the
global or parent environments. Make sure to namespace all function calls with
::. If additional R objects must be available in the function, you can
optionally supply a carrier::crate() with packaged data. It can also be NULL
for stores that don't need to embed their texts, for example, if only using
FTS algorithms such as ragnar_retrieve_bm25().}

\item{name}{Store name.}

\item{overwrite}{Whether to overwrite a preexisting store.}
}
\value{
String with the path to the store.
}
\description{
Create a store
}
\examples{
\donttest{
storeCreate(
  embed = \(x) ragnar::embed_ollama(x, model = "mxbai-embed-large"),
  name = "my_omop_assistant"
)

chat <- ellmerChat(
  name = "my_omop_assistant",
  chat = ellmer::chat_google_gemini(),
  top_k = 10L
)

chat$chat("How to create an acetaminophen cohort?")
}

}
