% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{storePath}
\alias{storePath}
\title{Path to a store}
\usage{
storePath(name = "omop_assistant.duckdb")
}
\arguments{
\item{name}{Store name.}
}
\value{
String with the path to the store.
}
\description{
Path to a store
}
\examples{
\donttest{
Sys.setenv("OMOP_DATA_FOLDER" = tempdir())
storePath()
}

}
