% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{trainModel}
\alias{trainModel}
\title{Create a store}
\usage{
trainModel(embed, chunks, name, overwrite = FALSE, ...)
}
\arguments{
\item{embed}{A function that is called with a character vector and returns a
matrix of embeddings. Note this function will be serialised and then
deserialised in new R sessions, so it cannot reference to any objects in the
global or parent environments. Make sure to namespace all function calls with
::. If additional R objects must be available in the function, you can
optionally supply a carrier::crate() with packaged data. It can also be NULL
for stores that don't need to embed their texts, for example, if only using
FTS algorithms such as ragnar_retrieve_bm25().}

\item{chunks}{a character vector or a dataframe with a text column, and
optionally, a pre-computed \code{embedding} matrix column. If \code{embedding} is not
present, then \code{store@embed()} is used. \code{chunks} can also be a character
vector.}

\item{name}{Store name.}

\item{overwrite}{Whether to overwrite a preexisting store.}

\item{...}{Passed to \code{ragnar::ragnar_store_create()}.}
}
\value{
Path to the trained model.
}
\description{
Create a store
}
