% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ellmerChat}
\alias{ellmerChat}
\title{Create an ellmer chat with the data of the omop trained model}
\usage{
ellmerChat(name = "omop_assistant", chat = ellmer::chat_google_gemini(), ...)
}
\arguments{
\item{name}{Store name.}

\item{chat}{A chat object.}

\item{...}{Arguments passed to \code{ragnar::ragnar_register_tool_retrieve()}}
}
\value{
The chat object with the trained store and prompt.
}
\description{
Create an ellmer chat with the data of the omop trained model
}
\examples{
\donttest{
storeCreate(
  embed = \(x) ragnar::embed_ollama(x, model = "mxbai-embed-large"),
  name = "my_omop_assistant"
)

chat <- ellmerChat(
  name = "my_omop_assistant",
  chat = ellmer::chat_google_gemini(),
  top_k = 10L
)

chat$chat("How to create an acetaminophen cohort?")
}

}
